+++ src/ConfigProvider.php
@@ -1,21 +1,21 @@
 <?php
+
 /**
- * @see       https://github.com/zendframework/zend-expressive-authentication-oauth2 for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authentication-oauth2/blob/master/LICENSE.md
- *     New BSD License
+ * @see       https://github.com/mezzio/mezzio-authentication-oauth2 for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Expressive\Authentication\OAuth2;
+namespace Mezzio\Authentication\OAuth2;
 
 use League\OAuth2\Server\AuthorizationServer;
-use League\OAuth2\Server\ResourceServer;
 use League\OAuth2\Server\Repositories\AccessTokenRepositoryInterface;
 use League\OAuth2\Server\Repositories\AuthCodeRepositoryInterface;
 use League\OAuth2\Server\Repositories\ClientRepositoryInterface;
 use League\OAuth2\Server\Repositories\RefreshTokenRepositoryInterface;
 use League\OAuth2\Server\Repositories\ScopeRepositoryInterface;
-use Zend\Expressive\Authentication\OAuth2\Repository\Pdo;
+use League\OAuth2\Server\ResourceServer;
+use Mezzio\Authentication\OAuth2\Repository\Pdo;
 
 class ConfigProvider
 {
@@ -44,7 +44,20 @@ class ConfigProvider
                 ClientRepositoryInterface::class => Pdo\ClientRepository,
                 RefreshTokenRepositoryInterface::class => Pdo\RefreshTokenRepository,
                 ScopeRepositoryInterface::class => Pdo\ScopeRepository,
-                UserRepositoryInterface::class => Pdo\UserRepository
+                UserRepositoryInterface::class => Pdo\UserRepository,
+
+                // Legacy Zend Framework aliases
+                \Zend\Expressive\Authentication\OAuth2\AuthCodeEntityInterface::class => AuthCodeEntityInterface::class,
+                \Zend\Expressive\Authentication\OAuth2\UserRepositoryInterface::class => UserRepositoryInterface::class,
+                \Zend\Expressive\Authentication\OAuth2\OAuth2Middleware::class => OAuth2Middleware::class,
+                \Zend\Expressive\Authentication\OAuth2\OAuth2Adapter::class => OAuth2Adapter::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\PdoService::class => Pdo\PdoService::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\AccessTokenRepository::class => Pdo\AccessTokenRepository::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\AuthCodeRepository::class => Pdo\AuthCodeRepository::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\ClientRepository::class => Pdo\ClientRepository::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\RefreshTokenRepository::class => Pdo\RefreshTokenRepository::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\ScopeRepository::class => Pdo\ScopeRepository::class,
+                \Zend\Expressive\Authentication\OAuth2\Repository\Pdo\UserRepository::class => Pdo\UserRepository::class,
             ],
             'factories' => [
                 OAuth2Middleware::class => OAuth2MiddlewareFactory::class,